+++ src/entities/packages/Health.cpp
@@ -11,7 +11,7 @@ void Health::set_max(float amount) {
 
 void Health::set_hp(float amount) { hp = amount; }
 
-void Health::set_invincibility(float amount) { invincibility_time = amount; }
+void Health::set_invincibility(float amount) { invincibility_time = static_cast<int>(amount); }
 
 void Health::update() {
 	hp = std::clamp(hp, 0.f, max_hp);
